feat: let users specify a memo when sending ft
[NEP-141](https://github.com/near/NEPs/blob/master/neps/nep-0141.md) specifies the
interface for fungible tokens.

The `ft_transfer` method has three parameters: `receiver_id`, `amount` and `memo`.
`memo` is an optional parameter.

This PR allows users to introduce a memo when using sending fungable tokens using
the cli.

The new parameter is optional and its skiped if no data is introduced.