feat: extended action builder with new released items (#457)
* Extended action builder for DeployGlobalContract and UseGlobalContract
actions introduced by nearcore-2.5.0 release
* Implemented todo related to transaction reconstruction

Please note that this PR creates 3 copies for each step in the action
builder.
Unfortunately, this is the only approach possible, and the difference
between files is the invocation of the next action step at the end.
I tried to create a loop, but it creates infinite recursion at compile
time, so this is the necessary solution.
That said, I didn't spend more than 10 minutes trying, so maybe there is
a way. I didn't want to spend much time trying to fix that as it's
implemented that way, and I think people before me tried and failed.

Example:

https://github.com/near/near-cli-rs/blob/440ef55c54f5ae63bac03c215b088d1c52370627/src/commands/transaction/construct_transaction/add_action_1/add_action/deploy_global_contract/mod.rs#L96,
https://github.com/near/near-cli-rs/blob/440ef55c54f5ae63bac03c215b088d1c52370627/src/commands/transaction/construct_transaction/add_action_2/add_action/deploy_global_contract/mod.rs#L96

To be honest, this action builder is sooo smelly. I believe we are
blocked here, but it is so sad to look at that. :(

```bash
# Constructing transactions using action builder:
./target/release/near transaction construct-transaction round-toad.testnet round-toad.testnet 
      add-action use-global-contract as-global-account-id yurtur2.near without-init-call
      add-action deploy-global-contract ./code.wasm as-global-account-id 
      add-action deploy-global-contract ./code.wasm as-global-hash
```

![image](https://github.com/user-attachments/assets/933ec8e5-e25b-4d0c-96a4-87aeb8b398c6)


@race-of-sloths